wip: worker logic implementation

+++ NetCommon/include/LogSearchCommon.hpp
@@ -17,11 +17,13 @@ namespace LogSystem {
         Server_JobFinished      // All work done
     };
 
+    // I chose 64 unsigned bit number over 32, 
+    // because 32 bit number alows to read maximum of 4GB file
     struct TaskPayload {
         char keyword[64];       // Search criteria
         char filename[128];
-        uint32_t start_line;    // Line offset
-        uint32_t end_line;      // End of chunk (0 means search till EOF)
+        uint64_t start_line;    // Line offset
+        uint64_t end_line;      // End of chunk (0 means search till EOF)
     };
 }
